+++ PresidentialGameEngine.ClassLibrary.Tests/CardComponentTests.cs
@@ -1,4 +1,5 @@
 ﻿using PresidentialGameEngine.ClassLibrary.Components;
+using PresidentialGameEngine.ClassLibrary.Enums;
 using static PresidentialGameEngine.ClassLibrary.Tests.TestStubsFakesAndMocks;
 
 namespace PresidentialGameEngine.ClassLibrary.Tests
