+++ PresidentialGameEngine.ClassLibrary/Enums/CardZone.cs
@@ -1,0 +1,11 @@
+﻿namespace PresidentialGameEngine.ClassLibrary.Enums
+{
+    public enum CardZone
+    {
+        Deck,
+        Hand,
+        Discard,
+        Removed,
+        CampaignStrategy
+    }
+}
