+++ PresidentialGameEngine.ClassLibrary/Exceptions/CardNotFoundException.cs
@@ -1,0 +1,8 @@
+﻿namespace PresidentialGameEngine.ClassLibrary.Exceptions
+{
+    public class CardNotFoundException : ArgumentException
+    {
+        public CardNotFoundException() : base() { }
+        public CardNotFoundException(string? message) : base(message) { }
+    }
+}
